Revert "Special lock"
