+++ simgui-ds.json
@@ -95,6 +95,7 @@
       "useGamepad": true
     },
     {
+      "guid": "78696e70757401000000000000000000",
       "useGamepad": true
     },
     {
